[BUGFIX] Remove vertical spacing of page layout grid
In order to streamline the spacing it is necessary that
all modules in the TYPO3 backend share the same base
stylings. The page layout grid currently adds unintentional
vertical spacing that is not present in any other component.
This space is now removed and the margin added to the
bottom is streamlined.

Resolves: #100575
Releases: main
Change-Id: I5a29bdbaf6f24c9ebc6957ac25025943f5490d74
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78614
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>