[TASK] Update table of submodule "Pages containing page TSconfig"
The naming of the table headers has been updated, as well as label
'Page title' in the pages TCA itself.

The spacing of the table cell containing the page title and icon has
been corrected to match the appearance of other tables containing
this information.

Also, the context menu has been added back and the page title link
has been removed.

Resolves: #100613
Releases: main
Change-Id: I404644f21acb23b3eea33636e9b979a2732c3666
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78657
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>