[TASK] Drop UserTS options `createFoldersInEB` and `folderTree.hideCreateFolder`
The UserTS options `createFoldersInEB` and `folderTree.hideCreateFolder`
were used in the past to control the existence of the "Create folder"
form in Element Browser instances. With the migration of the
"Create folder" view into a separate modal used in EXT:filelist, which
is based on Element Browser as well, those options became useless and
are therefore dropped.

Resolves: #100658
Releases: main
Change-Id: I100dceeffa262db8f7211bb9678d3162b2b16677
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78728
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>