[BUGFIX] Properly align positioning line at the end of tree
The positioning line is always shown above elements, regardless
of dragging position. This is most likely done to prevent jitter
of the line between elements.

This looks wrong at the end of the tree when trying to add a page
below the last child, so we add a condition to show the line at
the bottom for the last child of the tree in that case.

Resolves: #100392
Releases: main, 12.4
Change-Id: Ib5ae4fc93aff565f9f54ac5af3e5736138198b98
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79025
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Torben Hansen <derhansen@gmail.com>