[BUGFIX] Remove left over TCA noTableWrapping handling
TCA property 'noTableWrapping' for type 'user'
has been marked as deprecated with v9 already.
The v10 deprecation removals missed a usage in
PaletteAndSingleContainer, though.

The patch removes the handling and backports
to v12 as well as pre-patch to a more important
change in this area that needs a v12 backport
as well.

Resolves: #101154
Related: #86163
Related: #87464
Releases: main, 12.4
Change-Id: I427df1ce935acf2e0f91b2966d680f7fdc290e17
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79484
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>