[TASK] Simplify FormEngine PaletteAndSingleContainer
As a pre-patch to resolve #97330, we simplify the
code flow in PaletteAndSingleContainer by avoiding
the fieldSetWrap() method and inlining it in a
simplified form to the calling methods. This makes
the purpose more clear and allows detail changes
to the generated HTML depending on generated palettes.

This is just a code-flow cleanup that does not
influence generated HTML.

Resolves: #101155
Related: #97330
Releases: main, 12.4
Change-Id: Ibe13f2496d1e3df661b3049490d0a42d26329a30
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79487
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>