[TASK] Split the content of the installer progress bar
In order to ensure the readability of the progress bar, the previous
content has been split up. Only the percentages are now displayed in
the progress bar. Information about which installation step the user
is currently at is now displayed above the progress bar.

In addition, an undefined variable and icon were removed and the form
buttons of the first step are now aligned in a row.

Resolves: #100987
Releases: main, 12.4
Change-Id: Idc43d11a5f4b9988c51cf97482dee1f48d8e21b5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79425
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>