[BUGFIX] Make list of webmount page ids really unique
The fix for #100750 aimed to apply `array_unique()` to the array
containing page ids used as webmounts for the backend search.
Unfortunately, the call was done only for the storage in the runtime
cache with the side effect that still the full array with possible
duplications is returned on first invocation.

The affected `array_unique()` call is now extracted into a separate line
right before storing the resulting array to cache, which results in
returning the same, unified array.

Resolves: #101200
Related: #100750
Releases: main, 12.4, 11.5
Change-Id: Idbe6f1639ec84737ab151d5ab2d18739736e2327
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79560
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>