[BUGFIX] Streamline new record button titles in recordlist
To streamline the titles a specific language label is added for
the recordlist, which uses the TCA table title configuration of
every table. This leads to the fact that every new record button
now includes the specific table title analogous to recordlist-heading-title.

Additonally, this also unifies the modal header of the new
content element wizard, which previously was triggered using
different labels through the backend.

Resolves: #99824
Releases: main
Change-Id: Ib5819f2ae6e8ba639ae1dfc7e035ed9590c6241b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79699
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Annett Jähnichen <mcmietz@web.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Annett Jähnichen <mcmietz@web.de>