[BUGFIX] Accept "00:00 01-01-1970" in FormEngine UI
The FormEngine validation now accepts "00:00 01-01-1970" as an entered
date, which evaluates to int 0. If string 0 is passed, the field is
supposed to be empty. Previously, a too simple negation check was in
place.

In the long run, the whole date validation handling needs a major
rewrite to not rely on timestamps anymore.

Resolves: #101258
Releases: main, 12.4
Change-Id: I90a976088e01c23443369eff76f4c6401463715e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79819
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>