[BUGFIX] Use 12/24 hour format in date picker depending on client
flatpicker uses the 12 hours format by default, offering an AM/PM
selector next to the time dropdowns. In some regions, using this setting
is rather unusual and may lead to user errors.

To relax the sitation, browser API is used to determine the time format
based on the client's language. Unfortunately, it seems browsers always
assume the format based on the language only and ignore potentially
overridden settings in the operating system.

Resolves: #101347
Releases: main, 12.4
Change-Id: I96f92ea6c7ce52bf30bc96c1b8f5e49804fd85ab
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79991
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>