[BUGFIX] Respect item group of custom page type in "New Pages" view
With the introduced group and sorting feature for TCA items it was ever
since possible to easily define the way items were displayed in select
drop-downs.

This feature, however, was not considered in the "New Pages" view. This
is the view, which allows to create multiple pages at once. For each
page it is possible to select the type. This select drop-down now
correctly groups and sorts custom page types. For Core types the order
was already correct like it is in the TCA definition.

To achieve this, the "group and sort" functionality needed to be
extracted into an own service. The NewPages template also needed some
adjustments, as group labels are already translated by the service.

Resolves: #101355
Releases: main
Change-Id: I9203085cc816186887048383d96b85cd3c6d3925
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80029
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>