[BUGFIX] Alert about deleting content elements with references
If deleting content elements which have references
to it, a confirm dialog is displayed asking if the
element should really be deleted. Usually, the
dialog message contains information about existing
references (if there are references).

This was not the case if the content element is deleted
by using the inline delete button in the page module
in TYPO3 v11.

This behaviour was already improved and streamlined
in TYPO3 v12.

This is a partial backport of functionality in patch
https://review.typo3.org/c/Packages/TYPO3.CMS/+/72001
to backport the most critical functionality
improvement.

Releases: 11.5
Resolves: #101414
Related: #101411
Related: #92670
Related: #94780
Change-Id: Ie3ecad8f38a862c652baf115719b207820865708
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80116
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>