[BUGFIX] Display preview image of media resource
While the backend already renders preview images of for example
YouTube videos in the media module, it doesn't show the preview image
anywhere else, for example in the page module and the Media elements
within a "Text and Media" content element.

This patch shows the generated preview image within the page module
and the "Text and Media" content element as well.

Resolves: #95454
Releases: main, 12.4, 11.5
Change-Id: I4e404cf41627d5ee9ac165ee46817a456dbecaa8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80167
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: HOJA M A <hoja@ymail.com>
Reviewed-by: HOJA M A <hoja@ymail.com>
Tested-by: Henning Liebe <h.liebe@neusta.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>