[TASK] Streamline upper "Content" area in Page module localizations
In the page module it is possible to display the content of the
languages side by side with the "Language Comparison" mode.
With different amounts of text in the content elements of each
language, the page TSconfig option `mod.web_layout.defLangBinding = 1`
can help to arrange the content in a more structured way.

If the `mod.web_layout.defLangBinding = 1` option was activated, the
first button for creating new content elements was displayed on a white
background. With this patch, the button is now displayed on a gray
background, analogous to the rest of the content area.

Resolves: #101536
Releases: main, 12.4
Change-Id: I88a47d8010d0520bea84f01f9c21408e0f3908f9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80320
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jasmina Lie�mann <minapokhalo+typo3@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jasmina Lie�mann <minapokhalo+typo3@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Annett J�hnichen <mcmietz@web.de>