[FEATURE] Introduce new PSR-14 `BeforeLiveSearchFormIsBuiltEvent`
This change introduces a new PSR-14 `BeforeLiveSearchFormIsBuiltEvent`
which is dispatched in the LiveSearch form action. This event allows
extension developer to add, change or remove hints to the live search
form or change the search demand object.

Furthermore, additional view data can be provided and used in a
overridden module action template. This avoids the need for using
the `XCLASS` technique to provide additional data.

Resolves: #99409
Releases: main
Change-Id: I66c28693fd84600a9dd63d1864bda4a940ac0d06
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80436
Tested-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>