[BUGFIX] Keep "Add media by URL" modal open on error
If the "Add media by URL" button is used in FormEngine, a modal appears
with an input field, allowing the user to enter a media URL.
If an error occurs, the modal now stays open and the error is rendered
via a notification. Previously, the modal was closed and another modal
was opened informing about the error.

Resolves: #101650
Releases: 12.4
Change-Id: I3b0852bb485e041e1d008a6842cff6de1c27bb98
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80508
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Lina Wolf <112@linawolf.de>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Lina Wolf <112@linawolf.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>