[TASK] Deprecate EMU::addPageTSConfig()
ExtensionManagementUtility::addPageTSConfig() has
been superseded in v12 with feature #96614 already,
which allows autoloading Configuration/page.tsconfig
files.

To further reduce ExtensionManagementUtility,
ext_localconf.php files and TYPO3_CONF_VARS, the
old way is now marked as deprecated.

The patch does that, documents the impact and
refactors a couple of tests to deal with this
as well.

Resolves: #101799
Related: #96614
Releases: main
Change-Id: I1e71ea2c9dbc2d305668036d92911a7704b29c4a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80758
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>