[BUGFIX] Apply `t3js-colorpicker-initialized` class to correct element
Since migrating the color picker to alwan in #101646, the
`t3js-colorpicker-initialized` class is applied to the input element to
determine whether the color picker has been initialized already.

However, checking the class is done on a different element, which is a
simple left-over from the migration that is now fixed.

Resolves: #101839
Related: #101646
Releases: main
Change-Id: I5823056725d3b90843e8f10f45eed7802a912f3b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80853
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>