[BUGFIX] Toggle checkmark when switching aspect ratio
The image cropper component uses buttons to let a backend user toggle
the currently active aspect ratio setting, indicated via an `active`
class and a checkmark icon. With the removal of Font Awesome in #97877,
some CSS was removed that was responsible for toggling the visibility of
the checkmark icons.

This patch adds the missing CSS again, including necessary adaptions to
Core's Icon API.

Resolves: #101788
Related: #97877
Releases: main, 12.4
Change-Id: I1daf428cafba3bc5f1e707c2763bf3f5dd6ba293
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80842
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>