[TASK] Unbundle CKEditor5
Unbundle CKEditor5 in order to be able to load plugins like "font" only
if actually configured, allowing TYPO3 core to ship *all* available
ckeditor5 modules without having to weigh in the downsides for large
modules that we do not want to ship for "every" installation.

All @ckeditor scoped npm packages provided by CKEditor5 are bundled into
an own small per-package bundle that is loaded (and linked to each
other) via our importmap infrastructure. This allows specific modules to
be loaded ondemand and has the additional advantage that plugin authors
can use modules via packages exports as suggested by upstream
documentation. This should lower the the barrier for writing custom
plugins.

The existing bundle @typo3/ckeditor5-bundle is still in
place and acts as wrapper for the split bundles. It is
deprecated and will be removed in TYPO3 v13.0.

Also worth noting is that terser is now used to
shrink the module size.

Resolves: #102099
Releases: main, 12.4
Change-Id: I873167a7cb5c04cb7ef34e943830879f3f78eaaa
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81369
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>