[TASK] Show original user on new line in ElementHistoryController
This patch improves the readability of the record history view
of an element which was edited in switchTo mode by showing the
original user always on a new line.

div tags are used instead of span tags, to reduce any dependency
on bootstrap classes for displaying an element as display:block.

Resolves: #102200
Related: #102091
Releases: main, 12.4
Change-Id: I873711d4be6651ff15ae82e0bd7cbc4c6bf16e26
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81473
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>