[TASK] Proper DI in FlexFormTools
The event dispatcher should always be injected
via DI, either directly or through makeInstance().
It should not be nullable.

Switch a couple of service consumers to get an
instance injected.

Note further patches will remove left over state
properties from FlexFormTools ensuring the service
is fully stateless in the end, which is pretty
much the case already, but not obvious.

Resolves: #102213
Releases: main
Change-Id: I9c6dc87156ce1cfd69b8fdbb636364c456e91b60
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81491
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Nikita Hovratov <nikita.h@live.de>