[TASK] Improve History/Undo module
The History/Undo module can be used to undo changes.

> General:
The heading of the module and links referring to it in their naming have
been standardized ("History/Undo").

The module settings now use the same look as in other modules
(e.g. "TypoScript" or "TSconfig").

> Changelog table:
The look of the changelog is now similar to that in the log module.
The entries are now grouped by day and less relevant or redundant
information such as “age” and “day” have been removed from the table.

The record type and record title are now displayed in the module heading
in addition to the table and UID.
Analogously, the data record title (which is more relevant for editors)
is now displayed in the table column first, followed by the information
about the table and UID in brackets. The name of the table column has
also been updated accordingly.

A previously rather hidden link for filtering the changelog entries to a
specific record was highlighted as a separate button and the buttons
were moved to the right side of the table.

Previously, only selected action types ("insert" and "delete") were
shown by name in the table alongside the actual field changes. With this
patch, all action types are now listed in a new "Action" column, so that
it is clear at a first glance what kind of change it is (possible types
are: "add", "modifiy", "move", " delete", "undelete", "stage change").

> Preview for rollback:
In the preview for a rollback ("MultipleDiff"), the spacing has been
adjusted slightly and the buttons have been rearranged. Additionally the
button to roll back a single field is now only displayed if there are
changes to multiple fields.

The button "Show differences" in the module settings to switch
visibility of changes (field name only vs. field name with highlighted
changes) did not work in the MultipleDiff view, so its functionality
was restored.

> Outlook:
The module requires further improvements. For example, in the
MultipleDiff view, not all types of actions are evaluated, resulting in
an empty table in this view and therefore no differences are
highlighted/described if, for example, it is a "move" action.

If the changelog is limited to a selected record, any related changes
to, for example, file references will not be displayed. This also
results in an empty table in the MultipleDiff view.

The idea is to completely rewrite the ElementHistoryController and only
then make further adjustments to the MultipleDiff view.

Resolves: #102226
Releases: main
Change-Id: I7eeaaaa3979b769d109b74c4e334ae864ec0a230
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81505
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Annett Jähnichen <mcmietz@web.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>