[BUGFIX] Verify Popover instances before usage
In AC tests it appears that popover instances are not available, either
because the DOM isn't ready yet, or the element got removed meanwhile,
leading to AC test failures.

As a safety net, popover instances are now checked for not being `null`,
and a warning is logged in the console.

Also, the `@typo3/backend/context-help` module now waits for a ready DOM
and only takes elements into consideration that existed on
initialization time.

Resolves: #102368
Releases: main, 12.4
Change-Id: Icffc6f7e144881ded56910d7cb6f75b80f9752e1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81725
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>