[TASK] Position alert container at the bottom of the screen
By having a high z-index on the #alert-container, notifications
appear above the main interface elements, including dropdowns
from the topbar. However, this may lead to a usability issue, as the
notifications could obstruct user interaction with dropdowns and
docheader buttons like 'Share', 'Refresh', etc.

This commit updates the CSS to position the #alert-container element
at the bottom of the screen. For small screens it will be centered,
while for larger positioned to the right side.

Resolves: #102331
Releases: main, 12.4
Change-Id: If55bc67ca14d00d77c5bc53f624334a19cf0365d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81897
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Tested-by: Josef Glatz <typo3@josefglatz.at>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>