[BUGFIX] Show correct old name when renaming files/folders
When a file or folder is rename()'ed, the name property
is updated immediately. The file list controller handling
a rename thus needs to park the old name in a variable
to render the correct 'before' name.

The patch fixes this and cleans up the class a bit.

Resolves: #102478
Releases: main, 12.4
Change-Id: I438fa351febc821242c401cb208e644617f2f090
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82041
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>