[BUGFIX] Properly set allowed fields for record download
Since #101859, the `setFields` property of the record
list, which contains the fields to show / download is
initialized in the start() method.

To enable the "all columns" option again, overwriting
of the property is now done after calling start().

Additionally, the module data call is removed as this
is done by record list already.

Resolves: #102538
Related: #101859
Releases: main, 12.4
Change-Id: I39d7348b4eaa848f012cf238abc54e3a0e5d2e31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82002
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>