[!!!][TASK] Clean up Context
The constructor of class Context is unfortunate: It
needs to be registered in DI container to be injectable.
If an extension uses GU::makeInstance() with manual
constructor arguments of this class, it could shortcut
the container lookup and would create an own instance
not registered in the container, which could lead to
a second instance when Context is injected. The
constructor is only used in tests, tests are adapted
and constructor removed with the patch.

Also, the two classes ContextAwareInterface and
ContextAwareTrait look as if they are general things
that can be used similar to the PSR-3 LoggerAwareInterface
which is handled directly in DI / makeInstance(), but in
fact it is just a detail of the core routing aspect
handling. In those cases, DI / makeInstance() can be used
directly. We drop both classes to avoid confusion. Note
this also makes AspectFactory stateless.

Resolves: #102645
Releases: main
Change-Id: Ice0d4063bcfc8ff0c267f1b8155eab576d7d6772
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82159
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>