[TASK] Notify user about forced module redirect
There might be valid reasons for a module
being (temporary) inaccessible for a user.
Since #102665 a module redirect is done
instead of throwing an exception, which
a user can not resolve.

However, this led to the fact that a user
got redirected without knowing the reason.
Therefore, a notification is now dispatched,
whenever a forced module redirect takes place.

Resolves: #102678
Related: #102665
Releases: main, 12.4
Change-Id: I3e3e6548c7d2ec65b1081f8bdfdab5a16ba56f31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82241
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>