[TASK] Prevent possible TypeError with custom login providers
Check for existence of the password field before using its value to
prevent a "read properties of null" TypeError.

Note: This does usually not happen for the default login provider, but
might be for custom providers, not using the password field the same way
as the default provider.

Resolves: #102693
Releases: main
Change-Id: Id1ef515c517eb3f95020b97eb7336e618fcc1d30
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82253
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>