[BUGFIX] Omit calling `getRecordWSOL` in `BU::getProcessedValue()`
Since #102616, itemsProcFunc config is resolved in
`BU::getLabelFromItemlist()`. This method therefore
was adjusted to take the record in question as argument,
passing it to the underlying `ItemProcessingService`,
which actually just requires the uid and pid.

Since the database row is not passed to
BU::getProcessedValue()`, the record was
fetched for each item manually via
`getRecordWSOL()`. This is not sufficient
and has a serious performance impact.

The functionality of resolving itemsProcFunc is
now still in place but `BU::getProcessedValue()`
does no longer fetch and pass the record, but
only the uid and pid. This will therefore still
enable most use cases, but does no longer slow
down the backend.

Resolves: #102698
Related: #102616
Related: #100855
Releases: main, 12.4
Change-Id: Ic52564fada13b47f5364df642acbf683538b0c21
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82264
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>