[TASK] Avoid hard-coded table `tt_content` in GridColumnItem
This patch replaces all hard-coded occurrences with a class property,
allowing to set the table name as a constructor argument.

The page view could be way more flexible if it would allow to render
grid column items of other inline tables the same way as for tt_content
inside previews. This is done at the back of the mind to enable easy
Content Blocks backend previews for inline children.

Resolves: #102792
Releases: main, 12.4
Change-Id: I29ee892f1e3dbcb56cd0988039d1996ebf7ba157
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82414
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>