[BUGFIX] Translate Content Type labels if retrieved from fallback
This may happen if the CType is not defined via TCA but rather through
page TSConfig or itemsProcFunc.

This is really an edge-case, which was found during experiments on
backend previews for something else than tt_content records. Still, the
above scenario may theoretically happen and can easily be reproduced.

Resolves: #102823
Releases: main, 12.4
Change-Id: I7ed54ef339578fc6646c4baeacae7db6fe4c1048
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85199
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>