[TASK] Reduce hard-coded `tt_content` table in StandardPreviewRenderer
The goal here is to have a generic PreviewRenderer, which is not
tt_content specific. The StandardContentPreviewRenderer could then
extend this generic renderer. This is a pre-patch to make the
refactoring easier in the next patch.

For this to work, GridColumnItem now exposes methods to retrieve the
table from the context, the type column and the record type.

Additionally, some drive-by refactorings have been done like early
return statements and stricter checks.

The unused variable $infoArr has been removed as well.

Resolves: #102825
Releases: main, 12.4
Change-Id: I38eab3c095e06b789915600e79b7831556a25ca5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82442
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>