[BUGFIX] Restore `Controller` PHP attribute
Instead of using a class alias for the `Controller`
PHP attribute, is the class now restored.

This allows to install TYPO3 even without enabling the
`typo3/class-alias-loader` composer plugin when the deprecated
`Controller` attribute is used by an extensions. We also adapt our
codebase to use the newer `AsController` attribute as an example
for extensions authors that copy code from the sourcetree.

Resolves: #102858
Related: #102631
Releases: 12.4
Change-Id: Ia12843e9c29dc9bd32c2500e4c1c7d22aba4352a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82507
Tested-by: Josef Glatz <typo3@josefglatz.at>
Reviewed-by: Josef Glatz <typo3@josefglatz.at>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <ben@bnf.dev>