[TASK] Add stored page id to LiveSearch search demand
In preparation for further improvements in the LiveSearch, the latest
used page id stored in the browser's session storage is now passed to
the LiveSearch search demands.

This patch is marked as a task, as the whole API is still internal,
allowing to backport this to v12 as well.

Resolves: #102894
Releases: main, 12.4
Change-Id: If957d3b5eadbfe9fe5eb967fe5e17a697995ee6b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82592
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Siepmann <coding@daniel-siepmann.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Siepmann <coding@daniel-siepmann.de>