[BUGFIX] Change file extension separator in element browser string
The element browser string (bparams), forwarding
different information between states is using the
pipe ("|") character as separator.

With #101231, this string got extended by also
respecting a "disallowed" file extension list,
next to the "allowed" list. Since this information
is stored as one part, an additional separator
";" was introduced.

However, this character now leads to problems,
depending on the used environment (see linked
references). To circumvent this, the separator
is changed to the "~" character for now, as
this is not a reserved character, according
to RFC3986.

In the long run, this string should anyways be
replaced by a more robust solution.

* https://github.com/traefik/traefik/issues/9164
* https://github.com/symfony/symfony/pull/51709
* https://datatracker.ietf.org/doc/html/rfc3986#section-2.2

Resolves: #102992
Related: #101231
Releases: main, 12.4
Change-Id: I05e3548305d7ff50bd6db94fd2236aba95e80edb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82726
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>