[BUGFIX] Create backend users via CLI respects TCA default options
This change introduces the use of the DataHandler instead of the
QueryBuilder when creating backend users via CLI command
`backend:user:create`.

This ensures that the user is initialized with TCA default options and
the creation and history of the new user is logged to the TYPO3 log.

Resolves: #102062
Releases: main, 12.4
Change-Id: I78f7093108d1df3c6d37ab7545911365d3c846bb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83688
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>