[TASK] Reduce TCA ctrl origUid usages
The database column defined by TCA ctrl 'origUid'
('t3_origuid') had a bumpy ride over the years.

It was mostly obsoleted in v8 with the introduction of
'translationSource' ('l10n_source'). A patch to fully
remove it with #88495 has been reverted with #88501,
leaving the field in limbo.

The only remaining "true" usage of 'origUid' is in an
inline relation scenario in DataMapProcessor when an
inline parent *is* localizable (TCA ctrl 'languageField'
and 'transOrigPointerField' set), but a child *is not*:
Due to the copy strategy of all inline children upon
localization of a parent record, the core needs to know
if a child exists already or needs to be created. 'origUid'
then stores the uid of the default language child. A couple
of tests refer to this scenario as "monoglot" children.

We could potentially get rid of that scenario by
postulating that an inline child has to be localizable
as well if the parent is, but that decision can be
taken later.

The patch does not touch this scenario, but looks at
'origUid' from a different perspective: If a table *is*
localizable, or is not used as inline child, then 'origUid'
is not needed. The patch thus removes TCA ctrl 'origUid'
from TCA tables, keeping it only in the test scenario
extensions 'test_irre_csv', 'test_irre_foreignfield' and
'test_irre_foreignfield_non_ws'.

The bulk of the patch is about removing 't3_origuid' from
the functional test .csv fixture files, while the most
important TCA changes are dropping ctrl 'origUid' from
table 'pages' and 'tt_content'.

Resolves: #103099
Related: #79856
Related: #88494
Related: #88495
Related: #88501
Releases: main
Change-Id: I622afde74294c389639974d312f121d0a6466b60
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82888
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>