[FEATURE] Provide full userdata in password recovery email in ext:backend
Currently, the variables `name`, `email`, `language`, `resetLink`
and `username` are available in the template of the password recovery
email for backend users. It is not possible to add custom user
related data from the `be_users` table to the password recovery email.

To enable the possibility to add custom user related data, this
change adds the new variable `userData` to the FluidEmail object,
which contains all userdata of the affected backend user.

Additionally, a superfluous `is_array` check has been removed,
since it is ensured, that the `$users` variable at this point
already is an array.

Resolves: #103147
Releases: main
Change-Id: I72eb8796a15f5dd9dbb8682555777915f880bc96
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83031
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>